+++ libintervalxt/intervalxt/detail/component.ipp
