+++ libintervalxt/intervalxt/detail/connection.ipp
