+++ libintervalxt/intervalxt/detail/dynamical_decomposition.ipp
