+++ libintervalxt/intervalxt/detail/maybe_connection.ipp
