+++ libintervalxt/test/Makefile.am
@@ -22,7 +22,7 @@ EXTRA_DIST += external/catch2/single_include/catch2
 
 @VALGRIND_CHECK_RULES@
 
-AM_CPPFLAGS = -I $(srcdir)/../src -I $(builddir)/../src -isystem $(srcdir)/external/cereal/include -isystem $(srcdir)/external/catch2/single_include/catch2/
+AM_CPPFLAGS = -I $(srcdir)/../ -I $(builddir)/../ -isystem $(srcdir)/external/cereal/include -isystem $(srcdir)/external/catch2/single_include/catch2/
 AM_LDFLAGS = $(builddir)/../src/libintervalxt.la -lgtest -lbenchmark -lrt -lpthread
 # we use gmpxx; its header is using gmp directly so we need to link it in as well
 AM_LDFLAGS += -lgmpxx -lgmp
